requirements.txt
python-dotenv
pandas 
numpy
pymongo
certifi
pymongo[srv]==4.7
scikit-learn
-e .